Further reductions from the codestream
Moving on to hacking the J2K codestream a bit smaller, basically removing stuff until 0xFF90 (first tile) from the multiple .j2k codestreams, since it repeats in multiple files. It is going to be like a dictionary method since I need to store the different "headers" and transmit an index instead... Should save a lot, hundreds of bytes per .j2k, considering we have hundreds of views it should be tens of kilobytes.